* commit 'bfdcb9941b248855fb1e61a40dd04e3f52bdb091': (1979 commits)
  add missing 'default geometry' to feature type mapping
  Handle iframe referer set by browser issue
  Add transforming of the extent geometry to proper coordinate system when getting user wms layer
  fix condition in handling capabilities when 'version' param is empty
  added only_label
  Set layer download permission for Guest user
  remove dollar sign from the comment to prevent stopping flyway process
  prevent stopping printout process if legend graphic is not found
  add missing columns to sql queries for map layers
  comment the test file
  Fix adding attributes in WPS process. Fix saving default geometry to merged feature.
  Remove lenient parameter, set default targetEPSG to 3067, add userlayer to oskari_gis_data table
  removed V1_13 and renamed V1_14 to V1_13
  Add V1_14__fix and revert V1_10 to the one that was executed.
  renamed sql script
  flyway to remove hierarchical and add old layerselector2 and layerselection2
  add missing triggers to db, set 'lenient' param to true for solving "Bursa-Wolf Parameters Required" issue
  add work-around to avoid error in the flyway script
  V1_10__change_printout_legend_config.sql edited online with Bitbucket
  V1_10__change_printout_legend_config.sql edited online with Bitbucket
  ...

+++ service-map/src/main/java/fi/nls/oskari/map/userowndata/GisDataService.java
@@ -94,7 +94,7 @@ public class GisDataService
 	}
 	
 	public UserGisData getUserGisData(long id, boolean includeSharings) {
-		UserGisData result = _domainService.find(id);
+		UserGisData result = _domainService.find((int)id);
 		if (includeSharings) {
 			List<SharingItem> items = _sharingService.GetSharings(ResourceType.LAYER, id);		
 			List<UserGisDataSharing> sharing = new Vector<UserGisDataSharing>();
